# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: 'Scientific Documents Manager'
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  '/api/document/{id}':
    get:
      summary: 'GET Documents By id'
      operationId: gETDocumentsById
      description: "Returns document details by it's id"
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        200:
          description: Success
          content:
            text/plain:
              schema:
                type: string
                example: "{\n     \"id\": 53,\n     \"user_id\": 6,\n     \"document_state\": \"Pending\",\n     \"publish_date\": \"2025-12-22 18:00:11\",\n     \"create_date\": \"2023-04-26 14:25:04\",\n     \"update_date\": null,\n     \"delete_date\": null,\n     \"document_type\": {\n         \"id\": 1,\n         \"description\": \"Artigo Cientifico\"\n     },\n     \"document_metadata\": [\n         {\n             \"id\": 355,\n             \"value\": \"quis\",\n             \"metadata_type\": {\n                 \"id\": 1,\n                 \"description\": \"Titulo\"\n             }\n         },\n         {\n             \"id\": 356,\n             \"value\": \"laudantium\",\n             \"metadata_type\": {\n                 \"id\": 2,\n                 \"description\": \"Abstract\"\n             }\n         }\n     ],\n     \"document_media\": {\n         \"id\": 67,\n         \"model_type\": \"App\\\\Domain\\\\Aggregates\\\\Document\\\\Document\",\n         \"model_id\": 53,\n         \"uuid\": \"c6dd8e46-e8c6-4041-bc7f-cf86dcc7d6be\",\n         \"collection_name\": \"default\",\n         \"name\": \"book\",\n         \"file_name\": \"book.pdf\",\n         \"mime_type\": \"text/plain\",\n         \"disk\": \"media\",\n         \"conversions_disk\": \"media\",\n         \"size\": 407608,\n         \"manipulations\": [],\n         \"custom_properties\": [],\n         \"generated_conversions\": [],\n         \"responsive_images\": [],\n         \"order_column\": 1,\n         \"created_at\": \"2023-04-26T14:25:04.000000Z\",\n         \"updated_at\": \"2023-04-26T14:25:04.000000Z\",\n         \"original_url\": \"/storage/67/book.pdf\",\n         \"preview_url\": \"\"\n     },\n     \"document_file\":\n         \"JVBERi0xLjINJeLjz9MNCjY2IDAgb2JqDTw8I\n          MyBdIA0vTCAyMjkxNTYgDS9FIDIwMzU3IA0vT\n          ICAgICAgICAgICAgICAgICAgICAgICAgICAgI\n          MDAwMDE2IDAwMDAwIG4NCjAwMDAwMDExNDggM\n          OTU3IDAwMDAwIG4NCjAwMDAwMDIxNjYgMDAwM\n          IDAwMDAwIG4NCjAwMDAwMDM5MDAgMDAwMDAgb\n          MDAwIG4NCjAwMDAwMDUwOTkgMDAwMDAgbg0==\"\n}"
      tags:
        - 'Document(Internal) Endpoints'
    put:
      summary: 'PUT Edit Document'
      operationId: pUTEditDocument
      description: 'Edits document.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Changes Submited'
                properties:
                  message:
                    type: string
                    example: 'Changes Submited'
        400:
          description: 'Multiple Abstracts'
          content:
            application/json:
              schema:
                type: object
                example:
                  error: "You're document CAN ONLY have ONE abstract!"
                properties:
                  error:
                    type: string
                    example: "You're document CAN ONLY have ONE abstract!"
        403:
          description: 'Not document owner nor Admin'
          content:
            application/json:
              schema:
                type: object
                example:
                  error: "You don't have authorization to edit this file"
                properties:
                  error:
                    type: string
                    example: "You don't have authorization to edit this file"
      tags:
        - 'Document(Internal) Endpoints'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                document_type_id:
                  type: integer
                  description: 'ID of the document (Full list of ids can be retreived with api/document_types).'
                  example: 1
                publish_date:
                  type: datetime
                  description: 'Publish date of the document.'
                  example: '2025-12-22 18:00:11'
                document_metadata:
                  type: array
                  description: 'Document metadata. MUST HAVE AT LEAST A TITLE (id=1)'
                  example:
                    -
                      value: quis
                      metadata_type:
                        id: 1
                        description: 'Title 1'
                    -
                      value: laudantium
                      metadata_type:
                        id: 2
                        description: 'abstract 1'
                  items:
                    type: object
    delete:
      summary: 'DELETE Document'
      operationId: dELETEDocument
      description: "HARD deletes document.\nDO NOT USE if not certain."
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: object
                example:
                  message: 'Document Deleted'
                properties:
                  message:
                    type: string
                    example: 'Document Deleted'
        400:
          description: 'Document ID doesnt exist'
          content:
            application/json:
              schema:
                type: object
                example:
                  error: "The document you're trying to delete doesn't exist"
                properties:
                  error:
                    type: string
                    example: "The document you're trying to delete doesn't exist"
        403:
          description: 'Not document owner nor Admin'
          content:
            application/json:
              schema:
                type: object
                example:
                  error: "You don't have authorization to delete this file"
                properties:
                  error:
                    type: string
                    example: "You don't have authorization to delete this file"
      tags:
        - 'Document(Internal) Endpoints'
    parameters:
      -
        in: path
        name: id
        description: 'The id of the document.'
        example: 6
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []